Separated constants and helper functions from control panel.
